The save locations works now!!! working code as well

+++ app/src/main/java/hu/bme/ait/wanderer/data/LocationLabelCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(tableName = "location_label_cross_ref", primaryKeys = ["locationId","label"])
 data class LocationLabelCrossRef(
-    val locationId: Int,
+    val locationId: Long,
     val label: Label
 )
